remove script entries
